cdk_python/requirements.txt
aws-cdk-lib>=2.196.1
constructs>=10.4.2
boto3>=1.34.69
requests>=2.31.0
python-dotenv>=1.0.1
pyspark>=3.5.0 